Move easy-to-move cf current-users reducers
