+++ src/frontend/packages/cloud-foundry/src/store/reducers/current-user-roles-reducer/current-user-base-cf-role.reducer.ts
@@ -1,5 +1,5 @@
-import { GetCurrentUserRelationsComplete } from '../../../../cloud-foundry/src/actions/permissions.actions';
-import { getDefaultEndpointRoles, IAllCfRolesState } from '../../types/current-user-roles.types';
+import { GetCurrentUserRelationsComplete } from '../../../actions/permissions.actions';
+import { getDefaultEndpointRoles, IAllCfRolesState } from '../../../../../store/src/types/current-user-roles.types';
 import { currentUserCFRolesReducer } from './current-user-cf-roles.reducer';
 
 export function currentUserBaseCFRolesReducer(state: IAllCfRolesState = {}, action: GetCurrentUserRelationsComplete): IAllCfRolesState {
