+++ src/frontend/packages/cloud-foundry/src/store/reducers/current-user-roles-reducer/current-user-cf-roles.reducer.ts
@@ -1,11 +1,11 @@
-import { GetCurrentUserRelationsComplete } from '../../../../cloud-foundry/src/actions/permissions.actions';
-import { APIResource } from '../../types/api.types';
-import { getDefaultEndpointRoles, ICfRolesState, IOrgsRoleState } from '../../types/current-user-roles.types';
-import { isOrgRelation, isSpaceRelation } from './current-user-reducer.helpers';
+import { GetCurrentUserRelationsComplete } from '../../../actions/permissions.actions';
+import { APIResource } from '../../../../../store/src/types/api.types';
+import { getDefaultEndpointRoles, ICfRolesState, IOrgsRoleState } from '../../../../../store/src/types/current-user-roles.types';
+import { isOrgRelation, isSpaceRelation } from '../../../../../store/src/reducers/current-user-roles-reducer/current-user-reducer.helpers';
 import { createOrgRoleStateState } from './current-user-roles-org.reducer';
 import { currentUserOrgRolesReducer } from './current-user-roles-orgs.reducer';
 import { currentUserSpaceRolesReducer } from './current-user-roles-spaces.reducer';
-import { ISpace } from '../../../../core/src/core/cf-api.types';
+import { ISpace } from '../../../../../core/src/core/cf-api.types';
 
 export function currentUserCFRolesReducer(
   state: ICfRolesState = getDefaultEndpointRoles(),
