+++ src/frontend/packages/cloud-foundry/src/store/reducers/current-user-roles-reducer/current-user-roles-org.reducer.ts
@@ -1,5 +1,5 @@
-import { UserRelationTypes } from '../../../../cloud-foundry/src/actions/permissions.actions';
-import { IOrgRoleState } from '../../types/current-user-roles.types';
+import { UserRelationTypes } from '../../../actions/permissions.actions';
+import { IOrgRoleState } from '../../../../../store/src/types/current-user-roles.types';
 
 export const defaultUserOrgRoleState: IOrgRoleState = {
   isManager: false,
