+++ src/frontend/packages/cloud-foundry/src/store/reducers/current-user-roles-reducer/current-user-roles-orgs.reducer.ts
@@ -1,6 +1,6 @@
-import { GetCurrentUserRelationsComplete } from '../../../../cloud-foundry/src/actions/permissions.actions';
-import { IOrgsRoleState } from '../../types/current-user-roles.types';
-import { addNewRoles, removeOldRoles } from './current-user-reducer.helpers';
+import { GetCurrentUserRelationsComplete } from '../../../actions/permissions.actions';
+import { IOrgsRoleState } from '../../../../../store/src/types/current-user-roles.types';
+import { addNewRoles, removeOldRoles } from '../../../../../store/src/reducers/current-user-roles-reducer/current-user-reducer.helpers';
 import { currentUserOrgRoleReducer } from './current-user-roles-org.reducer';
 
 export function currentUserOrgRolesReducer(state: IOrgsRoleState = {}, action: GetCurrentUserRelationsComplete) {
