+++ src/frontend/packages/cloud-foundry/src/store/reducers/current-user-roles-reducer/current-user-roles-spaces.reducer.ts
@@ -1,6 +1,6 @@
-import { GetCurrentUserRelationsComplete } from '../../../../cloud-foundry/src/actions/permissions.actions';
-import { ISpacesRoleState } from '../../types/current-user-roles.types';
-import { addNewRoles, removeOldRoles } from './current-user-reducer.helpers';
+import { GetCurrentUserRelationsComplete } from '../../../actions/permissions.actions';
+import { ISpacesRoleState } from '../../../../../store/src/types/current-user-roles.types';
+import { addNewRoles, removeOldRoles } from '../../../../../store/src/reducers/current-user-roles-reducer/current-user-reducer.helpers';
 import { currentUserSpaceRoleReducer } from './current-user-roles-space.reducer';
 
 export function currentUserSpaceRolesReducer(state: ISpacesRoleState = {}, action: GetCurrentUserRelationsComplete) {
